fix app freeze on close pitchTap, add base class and send  to main thread async with lock
